one finder to rule them all
