--- development/tools/cmd/containerdoomfinder/go.mod
@@ -1,16 +1,0 @@
-module github.com/kyma-project/test-infra/development/tools/cmd/containerdoomfinder
-
-go 1.14
-
-require (
-	github.com/containerd/containerd v1.5.0-beta.4
-	github.com/containerd/continuity v0.0.0-20210315143101-93e15499afd5 // indirect
-	github.com/containerd/fifo v0.0.0-20210331061852-650e8a8a179d // indirect
-	github.com/gogo/googleapis v1.4.1 // indirect
-	github.com/google/go-cmp v0.5.5 // indirect
-	github.com/google/uuid v1.2.0 // indirect
-	github.com/sirupsen/logrus v1.8.1
-	golang.org/x/net v0.0.0-20210405180319-a5a99cb37ef4 // indirect
-	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
-	google.golang.org/grpc v1.36.1 // indirect
-)
