--- development/tools/cmd/dockeroomfinder/go.mod
@@ -1,8 +1,0 @@
-module github.com/kyma-project/test-infra/development/tools/cmd/dockeroomfinder
-
-go 1.14
-
-require (
-	github.com/fsouza/go-dockerclient v1.7.2
-	github.com/sirupsen/logrus v1.8.1
-)
