+++ development/tools/cmd/oomfinder/go.mod
@@ -1,0 +1,15 @@
+module github.com/kyma-project/test-infra/development/tools/cmd/dockeroomfinder
+
+go 1.14
+
+require (
+	github.com/containerd/containerd v1.4.3
+	github.com/docker/go-events v0.0.0-20190806004212-e31b211e4f1c // indirect
+	github.com/fsouza/go-dockerclient v1.7.2
+	github.com/gogo/googleapis v1.4.1 // indirect
+	github.com/google/uuid v1.2.0 // indirect
+	github.com/opencontainers/selinux v1.8.0 // indirect
+	github.com/sirupsen/logrus v1.8.1
+	github.com/syndtr/gocapability v0.0.0-20200815063812-42c35b437635 // indirect
+	google.golang.org/grpc v1.37.0 // indirect
+)
